+++ crates/radicle-protocol/src/service.rs
@@ -1177,6 +1177,7 @@ where
         match result {
             Ok(crate::worker::fetch::FetchResult {
                 updated,
+                canonical,
                 namespaces,
                 clone,
                 doc,
@@ -1198,6 +1199,16 @@ where
                     rid,
                     updated: updated.clone(),
                 });
+                self.emitter.emit_all(
+                    canonical
+                        .into_iter()
+                        .map(|(refname, target)| Event::CanonicalRefUpdated {
+                            rid,
+                            refname,
+                            target,
+                        })
+                        .collect(),
+                );
 
                 // Announce our new inventory if this fetch was a full clone.
                 // Only update and announce inventory for public repositories.