+++ crates/radicle/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Introduce a node event for canonical reference updates, `Event::CanonicalRefUpdated`.
+  Whenever the node fetches new updates, it checks if canonical references can
+  be updated. The node has learned how to return these results and emit them as
+  node events. This is a breaking change since it adds a new variant the `Event`
+  type.
+- Add `#[non_exhaustive]` to `Event` to prevent any further breaking changes
+  when adding new variants.
+
 ### Changed
 
 - `radicle::profile::Home::socket` defaults to the path `\\.\pipe\radicle-node`